Site: Re-organized app directory structure; fixes #1077
