+++ src/main/webapp/site/src/app/app.module.ts
@@ -2,13 +2,13 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { HeaderModule } from './header/header.module';
-import { FooterModule } from './footer/footer.module';
+import { HeaderModule } from './modules/header/header.module';
+import { FooterModule } from './modules/footer/footer.module';
 import { StudentModule } from './student/student.module';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { StudentService } from './student/student.service';
-import { UserService } from './user.service';
+import { UserService } from './services/user.service';
 import { HttpClientModule } from '@angular/common/http';
