+++ src/main/webapp/site/src/app/modules/footer/footer.component.spec.ts
