+++ src/main/webapp/site/src/app/modules/header/header-account-menu/header-account-menu.component.scss
