+++ src/main/webapp/site/src/app/modules/header/header-account-menu/header-account-menu.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { User } from '../../user';
+import { User } from "../../../domain/user";
 
 @Component({
   selector: 'app-header-account-menu',
