+++ src/main/webapp/site/src/app/modules/header/header.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { User } from '../user';
-import { UserService } from "../user.service";
+import { User } from '../../domain/user';
+import { UserService } from "../../services/user.service";
 
 @Component({
   selector: 'app-header',
