+++ src/main/webapp/site/src/app/modules/shared/search-bar/search-bar.component.ts
