+++ src/main/webapp/site/src/app/modules/shared/select-menu/select-menu.component.scss
