+++ src/main/webapp/site/src/app/modules/shared/shared.module.ts
