+++ src/main/webapp/site/src/app/student/student-run-list/student-run-list.component.spec.ts
@@ -7,8 +7,8 @@ import { StudentService } from '../student.service';
 import { StudentRunListComponent } from './student-run-list.component';
 import { StudentRunListItemComponent } from '../student-run-list-item/student-run-list-item.component';
 import { MatCardModule, MatIconModule, MatFormFieldModule, MatInputModule, MatSelectModule } from "@angular/material";
-import { SelectMenuComponent } from "../../shared/select-menu/select-menu.component";
-import { SearchBarComponent } from "../../shared/search-bar/search-bar.component";
+import { SelectMenuComponent } from "../../modules/shared/select-menu/select-menu.component";
+import { SearchBarComponent } from "../../modules/shared/search-bar/search-bar.component";
 import { FormsModule } from "@angular/forms";
 import { MomentModule } from "angular2-moment";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
